Move all sending signals to the fitting `sending` proto files.
