+++ model/src/main/proto/spine_examples/chatspn/message/commands.proto
@@ -52,19 +52,3 @@ message PostMessage {
     // The message text content.
     string content = 4 [(required) = true];
 }
-
-// Tells to update a message content.
-message UpdateMessageContent {
-
-    // The ID of the message to update its content.
-    MessageId id = 1;
-
-    // The ID of the chat to update message content in.
-    ChatId chat = 2 [(required) = true];
-
-    // The ID of the user who tells to update message content.
-    spine.core.UserId user = 3 [(required) = true];
-
-    // The suggested message content.
-    string suggested_content = 4 [(required) = true];
-}
